-remove shimmer functions from design system -remove unused dependencies from gradle
