+++ presentation/build.gradle.kts
@@ -52,7 +52,6 @@ dependencies {
 
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.work.runtime.ktx)
-    implementation(libs.androidx.compose.material)
     debugImplementation(libs.ui.tooling)
 
     implementation(libs.foundation)
